requirements.txt
asgiref==3.6.0
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.11.2
certifi==2022.12.7
charset-normalizer==3.0.1
contourpy==1.0.7
cycler==0.11.0
Django==4.1.7
exceptiongroup==1.1.0
fake-useragent==1.1.1
fonttools==4.38.0
h11==0.14.0
idna==3.4
kiwisolver==1.4.4
lxml==4.9.2
matplotlib==3.7.0
numpy==1.24.2
opencv-python==4.7.0.72
outcome==1.2.0
packaging==23.0
Pillow==9.4.0
psycopg2-binary==2.9.5
pyparsing==3.0.9
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
response==0.5.0
scipy==1.10.1
selenium==4.8.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
sqlparse==0.4.3
trio==0.22.0
trio-websocket==0.9.2
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
wsproto==1.2.0
